test: test all endpoints with auth
